pyproject.toml
[tool.poetry]
name = "dialbot"
version = "0.0.2"
description = ""
authors = ["Stephen Stack <sfstack7500@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.1"
openai = "^1.11.1"
gunicorn = "^21.2.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
